Financial-Sentiment-Analysis/requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

# NLP & Sentiment Analysis
nltk>=3.8.1
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
vaderSentiment>=3.3.2

# Data Collection & APIs
requests>=2.31.0
tweepy>=4.14.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
newsapi-python>=0.2.7
yfinance>=0.2.28
alpha-vantage>=2.3.1

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework (Optional - for API)
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
schedule>=1.2.0

# Development & Testing
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
ipython>=8.12.0
ipykernel>=6.25.0

# Logging & Monitoring
loguru>=0.7.0
